+++ src/propertyFields/collectionData/collectionDataViewer/ICollectionDataViewerState.ts
@@ -3,5 +3,6 @@ import { FieldValidator } from "..";
 export interface ICollectionDataViewerState {
   crntItems: any[];
   inCreationItem?: any;
+  inCreationItemValid?: boolean;
   validation?: FieldValidator;
 }