+++ src/webparts/propertyControlsTest/PropertyControlsTestWebPart.ts
@@ -675,7 +675,8 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                   multiSelect: true,
                   onPropertyChange: this.onPropertyPaneFieldChanged,
                   properties: this.properties,
-                  key: 'sitesFieldId'
+                  key: 'sitesFieldId',
+                  trimDuplicates: true
                 }),
                 PropertyFieldTeamPicker('teams', {
                   key: 'teamsPicker',
@@ -1025,6 +1026,7 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                   disabled: false,
                   key: 'codeEditorFieldId',
                   language: PropertyFieldCodeEditorLanguages.HTML,
+                  panelWidth: '700px'
                   // options: {
                   //   wrap: true,
                   //   fontSize: 20,
@@ -1059,7 +1061,10 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                     {
                       id: "Lastname",
                       title: "Lastname",
-                      type: CustomCollectionFieldType.string
+                      type: CustomCollectionFieldType.string,
+                      onGetErrorMessage: (value, index, currentItem) => {
+                        return value == 'Smith' && currentItem.City == 'antwerp' ? 'You cannot write Smith when City is Antwerp' : ""; 
+                      }
                     },
                     {
                       id: "Age",
@@ -1086,6 +1091,14 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                         {
                           key: "montreal",
                           text: "Montreal"
+                        },
+                        {
+                          key: false,
+                          text: 'False City'
+                        },
+                        {
+                          key: 0,
+                          text: 'Zero City'
                         }
                       ],
                       required: true,
@@ -1103,7 +1116,10 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                       id: "Sign",
                       title: "Signed",
                       type: CustomCollectionFieldType.boolean,
-                      defaultValue: true
+                      defaultValue: true,
+                      onGetErrorMessage: (value, index, currentItem) => {
+                        return value && currentItem.City == 'antwerp' ? 'You cannot check sign when City is Antwerp' : "";
+                      }
                     },
                     {
                       id: "IconName",
@@ -1124,7 +1140,10 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                       id: "color",
                       title: "Color",
                       type: CustomCollectionFieldType.color,
-                      defaultValue: "#ff0000"
+                      defaultValue: "#ff0000",
+                      onGetErrorMessage: (value, index, currentItem) => {
+                        return value == '#ff0000' && currentItem.City == 'antwerp' ? 'You cannot set default color when City is Antwerp' : ""; 
+                      }
                     },
                     {
                       id: "custom",
@@ -1145,6 +1164,18 @@ export default class PropertyControlsTestWebPart extends BaseClientSideWebPart<I
                             }), " 🎉"
                           )
                         );
+                      },
+                      onGetErrorMessage: (value, index, currentItem) => {
+                        return value == 'hello' && currentItem.City == 'antwerp' ? 'You cannot write hello when City is Antwerp' : ""; 
+                      }
+                    },
+                    {
+                      id: "customVisibility",
+                      title: "Custom Visible Field",
+                      type: CustomCollectionFieldType.boolean,
+                      defaultValue: true,
+                      isVisible: (field, items) => {
+                        return items.filter(i => i.City == 'antwerp').length > 0;
                       }
                     }
                   ],